requirements.txt
asgiref==3.8.1
boto3==1.35.28
botocore==1.35.28
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
distlib==0.3.8
Django==5.0.6
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
gunicorn==23.0.0
ibmcloud==0.1.0
idna==3.10
jmespath==1.0.1
packaging==24.2
paypalrestsdk==1.13.3
pillow==10.3.0
platformdirs==4.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.10.2
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tornado==4.5.2
typing_extensions==4.13.2
urllib3==2.2.1
virtualenv==20.26.3
whitenoise==6.8.2
